requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai
redis==5.0.1
qdrant-client==1.14.3
PyPDF2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.8.5
psycopg[binary]==3.2.9
python-multipart
